Execute commands where single quotes are needed.
This is in response to issue #32 where the user wanted to enter a
command where they specified the shell. eg. `bash -c 'echo testing'`.

This fix will not handle a command with trailing `&&` at this time but
after looking at the readme it seems that may not be the proper use case
anyway. eg. `bash -c 'echo testing' && echo again`