Bug fixes and improvements for corner cases
Fix the defect that Column crashes on zero or negative input.
Fix the defect that Last() crashes on negative input;
Fix the defect that Regex filters crashes on nil regex expression;
Add unit test for WithError(); close the pipe when a non-nil error is set;
Add ExecForEach, Reject(), and RejectRegexp to doMethodsOnPipe tests;
Improve coverage by adding equal counts in the testcase of Freq;
Use io.Copy() in Stdout() to improve efficiency.