Fix Bytes/AppendFile/WriteFile to return read errs
Calling any sink method (such as Bytes) should return a non-nil error if there is an error reading the pipe to completion. Due to a bug, Bytes, AppendFile, and WriteFile failed to do this. This commit fixes the bug, and also tests that the other sink methods behave correctly (which they do). Fixes #170.

Co-authored-by: Namho Kim <git@namho.kim>