Remove arbitrary 64KiB limit on line length
The default bufio.Scanner is limited to 64 * 1024 bytes, throwing a 'token too
long' error if a line exceeds this size. This commit sets the token limit
instead to math.MaxInt, which is the biggest possible slice capacity in Go, and
effectively limits the scan line length only by available memory (fixes #167).

Co-authored-by: Tim Hobbs <timothy.hobbs@servicelayers.com>